requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
pydantic>=2.7
PyYAML>=6.0
structlog>=24.1
tenacity>=8.2
slowapi>=0.1.9
pytest>=8.2
pytest-asyncio>=0.23
anyio>=4.3

